pyproject.toml
[project]
name = "espimg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.3",
    "pillow>=11.0.0",
]
